moved frame sampling to be a filter
