+++ pliers/filters/__init__.py
@@ -1,5 +1,9 @@
 ''' Filter hierarchy. '''
 
 from .text import WordStemmingFilter
+from .video import FrameSamplingFilter
 
-__all__ = ['WordStemmingFilter']
+__all__ = [
+    'WordStemmingFilter',
+    'FrameSamplingFilter'
+]
